requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.17
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
eval_type_backport==0.2.2
filelock==3.18.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
frozenlist==1.6.0
fsspec==2025.3.2
greenlet==3.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
lxml==5.3.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.4.3
networkx==3.4.2
nltk==3.9.1
numpy==2.2.5
openai==1.76.0
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
propcache==0.3.1
pyarrow==19.0.1
PyAutoGUI==0.9.54
pydantic==2.11.3
pydantic_core==2.33.1
PyGetWindow==0.0.9
Pygments==2.19.1
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-pptx==1.0.2
pytweening==1.2.0
pytz==2025.2
pywhatkit==5.4
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.3
setuptools==79.0.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
sympy==1.13.3
tabulate==0.9.0
together==1.5.5
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.51.3
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
wikipedia==1.4.0
WTForms==3.2.1
XlsxWriter==3.2.3
yarl==1.20.0
supabase==2.6.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
